% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhs.r
\name{inRegion}
\alias{inRegion}
\title{Subfunction for \link{fhs}}
\usage{
inRegion(data, v)
}
\arguments{
\item{data}{Vector of (x,y)}

\item{v}{Data frame of ordered vertices with (x,y) co-ordinates in rows}
}
\value{
Returns logical of whether point is in polygon
}
\description{
Subfunction
}
\details{
Subfunction
}
\author{
Wajid Jawaid
}

